requirements.txt
easyocr==1.6.2
Flask==2.2.2
Jinja2==3.1.2
imageio==2.28.1
importlib-metadata==6.6.0
importlib-resources==5.12.0
matplotlib==3.7.1
numpy
opencv-python
opencv-python-headless
opencv-contrib-python==4.5.5.62
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
pyparsing==3.0.9
requests==2.30.0
tensorflow==2.12.0
thop==0.1.1.post2209072238
psutil==5.9.5
torch==2.0.1
torchvision==0.15.2
ultralytics
Werkzeug==2.3.4
zipp==3.15.0
pip~=23.1.2
attrs~=23.1.0
wheel~=0.40.0
utils~=1.0.1
setuptools~=67.7.2
flask-cors~=3.0.10